change height unit from u-int to CGFloat

+++ CalendarView/CCLProvidesDetailView.h
@@ -17,5 +17,5 @@
 - (NSView *)detailViewForObjectValue:(id)objectValue;
 
 @optional
-- (NSUInteger)heightOfDetailView;
+- (CGFloat)heightOfDetailView;
 @end
